move the label util to use in metadata
