+++ pkg/api/metadata.go
@@ -9,6 +9,8 @@ import (
 	"strings"
 
 	"github.com/sirupsen/logrus"
+
+	"github.com/openshift/ci-tools/pkg/api/utils"
 )
 
 // IsComplete returns an error if at least one of Org, Repo, Branch members is
@@ -143,11 +145,7 @@ func BuildCacheFor(metadata Metadata) ImageStreamTagReference {
 }
 
 func ImageVersionLabel(fromTag PipelineImageStreamTagReference) string {
-	label := fmt.Sprintf("io.openshift.ci.from.%s", fromTag)
-	if len(label) > 63 { // The max length of a label is 63 chars
-		label = fmt.Sprintf("%s-cont", label[:58])
-	}
-	return label
+	return utils.Trim63(fmt.Sprintf("io.openshift.ci.from.%s", fromTag))
 }
 
 var testPathRegex = regexp.MustCompile(`(?P<org>[^/]+)/(?P<repo>[^@]+)@(?P<branch>[^:]+):(?P<test>.+)`)
